refactor: 게시물 추가,  comment 로직 수정
