+++ package.json
@@ -13,11 +13,12 @@
   },
   "dependencies": {
     "@tanstack/react-query": "^5.84.2",
+    "@tanstack/react-query-devtools": "^5.85.3",
+    "overlay-kit": "^1.4.1",
     "react": "^19.1.1",
     "react-dom": "^19.1.1",
-    "zustand": "^5.0.7",
     "react-error-boundary": "^6.0.0",
-    "overlay-kit": "^1.4.1"
+    "zustand": "^5.0.7"
   },
   "devDependencies": {
     "@eslint/js": "^9.33.0",
